refactor: 이벤트 구조 변ㅕㅇ
