+++ src/main/java/org/ormi/stackorflow/infra/message/DiscordMessageEntity.java
@@ -1,0 +1,23 @@
+package org.ormi.stackorflow.infra.message;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class DiscordMessageEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
+
+  @Column(name = "discord_user_id", updatable = false)
+  private long discordUserId;
+
+
+}
