+++ src/main/java/org/ormi/stackorflow/infra/modules/discord/DiscordConfiguration.java
@@ -2,9 +2,7 @@ package org.ormi.stackorflow.infra.modules.discord;
 
 import discord4j.core.DiscordClientBuilder;
 import discord4j.core.GatewayDiscordClient;
-import discord4j.core.event.domain.Event;
 import java.util.Objects;
-import java.util.Set;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -12,22 +10,12 @@ import org.springframework.context.annotation.Configuration;
 public class DiscordConfiguration {
 
   @Bean
-  public <T extends Event> GatewayDiscordClient discordClient(
-      DiscordBotToken discordBotToken,
-      Set<DiscordListener<T>> discordListeners
-  ) {
+  public GatewayDiscordClient discordClient(DiscordBotToken discordBotToken) {
     GatewayDiscordClient discordClient = DiscordClientBuilder.create(discordBotToken.getBotToken())
         .build()
         .login()
         .block();
     Objects.requireNonNull(discordClient, "discordClient");
-
-    discordListeners.forEach(listener -> {
-      discordClient.on(listener.type())
-          .flatMap(listener::execute)
-          .onErrorResume(listener::handleError)
-          .subscribe();
-    });
 
     return discordClient;
   }
