+++ src/main/java/org/ormi/stackorflow/infra/modules/discord/DiscordMessageCreatedEventListener.java
@@ -1,6 +1,10 @@
 package org.ormi.stackorflow.infra.modules.discord;
 
+import discord4j.core.GatewayDiscordClient;
 import discord4j.core.event.domain.message.MessageCreateEvent;
+import discord4j.core.object.entity.channel.Channel;
+import discord4j.discordjson.json.ChannelData;
+import discord4j.rest.entity.RestChannel;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
@@ -11,9 +15,14 @@ import reactor.core.publisher.Mono;
 final class DiscordMessageCreatedEventListener implements DiscordListener<MessageCreateEvent> {
 
   private final ApplicationEventPublisher eventPublisher;
+  private final GatewayDiscordClient discordClient;
 
-  public DiscordMessageCreatedEventListener(ApplicationEventPublisher eventPublisher) {
+  public DiscordMessageCreatedEventListener(
+      ApplicationEventPublisher eventPublisher,
+      GatewayDiscordClient discordClient
+  ) {
     this.eventPublisher = eventPublisher;
+    this.discordClient = discordClient;
   }
 
   @Override
@@ -21,8 +30,19 @@ final class DiscordMessageCreatedEventListener implements DiscordListener<Messag
     if (event.getMember().isEmpty()) {
       return Mono.empty();
     }
-    DiscordMessageCreatedEvent applicationEvent = DiscordMessageCreatedEvent.from(event);
-    eventPublisher.publishEvent(applicationEvent);
+    Mono<ChannelData> channelDataMono = discordClient.getChannelById(event.getMessage().getChannelId())
+      .map(Channel::getRestChannel)
+      .flatMap(RestChannel::getData);
+    Mono.zip(
+        Mono.just(DiscordCreatedMessage.from(event)),
+        channelDataMono
+    )
+    .map(tuple -> {
+      DiscordChannel discordChannel = DiscordChannel.from(tuple.getT2());
+
+      return DiscordMessageCreatedEvent.from(discordChannel,tuple.getT1());
+    })
+    .subscribe(eventPublisher::publishEvent);
 
     return Mono.empty();
   }
